+++ files_service.h
@@ -9,9 +9,9 @@
 
 using namespace std;
 
-vector<int> read_file(string path);
+vector<uint8_t> read_file(string path);
 vector<string> exec(string cmd);
 vector<string> file_names(string path, string extension);
-vector<vector<int>> read_all_files(vector<string> list_files, string part_of_path);
+vector<vector<uint8_t>> read_all_files(vector<string> list_files, string part_of_path);
 
 #endif
