+++ gpuLROO.h
@@ -8,8 +8,8 @@
 using namespace std;
 
 __device__ int gpu_specify_type(int n);
-__device__ double gpuLongestRunOfOnes(int n, int data[]);
-__global__ void gpu(int *a, double *b, int z, int n);
+__device__ double gpuLongestRunOfOnes(int n, uint8_t data[]);
+__global__ void gpu(uint8_t *a, double *b, int z, int n);
 
 
 #endif
